Major addition: Added a shell script that runs the simulation with different settings (set in a text file). The output from the initial run of this file has also been added for further analysis.
Other changes :
Added a small perturbation to pref velocity passed to velocity calc to avoid problems caused by perfect symmetry in RVO2. Too small to affect the other models.

Chart rendering implemented in RVOGui to ensure that charts are not created if simulation is run without GUI.

RVOModel changed so that data is written to file each time the simulation is stopped. (previously used to write only when restarted)