Changed how the roadmap is implemented. now each roadmap point has an associated integer which indicates it's position in the sequence of roadmap points. if two points are oart if a decision to be made by the agent then the agent decides to move to the closest of the points with the same integer value.
Changed the value of obst horizon for rvo2 in lattice test settings to work better in narrow corridors

added a new roadMap.xml in EvacTest folder which demonstrates and tests how the new roadMap works.

As of now the creator GUI does not allow you to specify the integer value of each road map point. This has to be done in the xml file by hand.