Fixed data collector .
Only velocity and position are written. in a folder and file of structure: "ScenarioNumber\MotionPlanner\Seed\ScenarioNumber_MotionPlanner_Seed"

Written as a binary file with each column corresponding to an agent and each row corresponds to a timestep

For the lattice model  , the lattice state is writen as a 2d int array into a binary file:  an empty line seperates one timestep from the next. A single space seperates each column from the next